PropertyDocumentReview/requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.9
lxml==6.0.2
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdfium2==5.1.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.1
reportlab==4.4.5
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
